# Conflicts:
#	.gitignore
#	frontend/package-lock.json
#	frontend/package.json
#	frontend/public/img/black_heart.png
#	frontend/public/img/heart.png
#	frontend/public/index.html
#	frontend/src/App.js
#	frontend/src/Main.js
#	frontend/src/asset/css/main.css
#	frontend/src/asset/css/user.css
#	frontend/src/component/DateRangeSelector.js
#	frontend/src/component/Footer.js
#	frontend/src/component/Header.js
#	frontend/src/component/HotelItem.js
#	frontend/src/component/HotelRoomsItem.js
#	frontend/src/component/KakaoMap.js
#	frontend/src/component/Reply.js
#	frontend/src/component/SliderReviewItems.js
#	frontend/src/pages/admin/aguest.js
#	frontend/src/pages/admin/ahost.js
#	frontend/src/pages/admin/amain.js
#	frontend/src/pages/admin/css/astyles.css
#	frontend/src/pages/admin/notice/alist.js
#	frontend/src/pages/admin/notice/awrite.js
#	frontend/src/pages/guest/Order.js
#	frontend/src/pages/guest/aa.css
#	frontend/src/pages/guest/modall.css
#	frontend/src/pages/guest/select.js
#	frontend/src/pages/host/HostPage.js
#	frontend/src/pages/host/HotelDetail.js
#	frontend/src/pages/host/host1.css
#	frontend/src/pages/host/hostAccount/EditHostInfo.js
#	frontend/src/pages/host/hostAccount/EditReply.js
#	frontend/src/pages/host/hostAccount/HostAccount.js
#	frontend/src/pages/host/hostAccount/ListReviews.js
#	frontend/src/pages/host/hostAccount/ReviewItem.js
#	frontend/src/pages/host/hostAccount/WriteReply.js
#	frontend/src/pages/host/hotelDetailSection/HostInfo.js
#	frontend/src/pages/host/hotelDetailSection/HotelAmenities.js
#	frontend/src/pages/host/hotelDetailSection/HotelDescription.js
#	frontend/src/pages/host/hotelDetailSection/HotelRooms.js
#	frontend/src/pages/host/hotelDetailSection/Reputation.js
#	frontend/src/pages/host/hotelDetailSection/ReputationItem.js
#	frontend/src/pages/host/hotelDetailSection/Reputation_modal.js
#	frontend/src/pages/host/hotelDetailSection/Reservation.js
#	frontend/src/pages/host/login/Join_modal.js
#	frontend/src/pages/host/login/Login.js
#	frontend/src/pages/host/login/SearchEmail.js
#	frontend/src/pages/host/login/SearchPw.js
#	syFinal/.gradle/8.7/checksums/checksums.lock
#	syFinal/build.gradle
#	syFinal/src/main/java/com/example/syFinal/MainController.java
#	syFinal/src/main/java/com/example/syFinal/admin/controller/AGuestController.java
#	syFinal/src/main/java/com/example/syFinal/admin/controller/AHostController.java
#	syFinal/src/main/java/com/example/syFinal/admin/controller/ALoginController.java
#	syFinal/src/main/java/com/example/syFinal/admin/controller/ANoticeController.java
#	syFinal/src/main/java/com/example/syFinal/admin/model/dao/AGuestDAO.java
#	syFinal/src/main/java/com/example/syFinal/admin/model/dao/AGuestDAOImpl.java
#	syFinal/src/main/java/com/example/syFinal/admin/model/dao/ANoticeDAO.java
#	syFinal/src/main/java/com/example/syFinal/admin/model/dao/ANoticeDAOImpl.java
#	syFinal/src/main/java/com/example/syFinal/admin/model/dao/AdminDAO.java
#	syFinal/src/main/java/com/example/syFinal/admin/model/dao/AdminDAOImpl.java
#	syFinal/src/main/java/com/example/syFinal/admin/model/dao/AhostDAO.java
#	syFinal/src/main/java/com/example/syFinal/admin/model/dao/AhostDAOImpl.java
#	syFinal/src/main/java/com/example/syFinal/admin/model/dto/AHostDTO.java
#	syFinal/src/main/java/com/example/syFinal/admin/model/dto/ANoticeDTO.java
#	syFinal/src/main/java/com/example/syFinal/admin/model/dto/AdminDTO.java
#	syFinal/src/main/java/com/example/syFinal/global/PageUtil.java
#	syFinal/src/main/java/com/example/syFinal/global/controller/ReputationController.java
#	syFinal/src/main/java/com/example/syFinal/global/model/EmailService.java
#	syFinal/src/main/java/com/example/syFinal/global/model/EmailServiceImpl.java
#	syFinal/src/main/java/com/example/syFinal/global/model/ReputationDAO.java
#	syFinal/src/main/java/com/example/syFinal/global/model/ReputationDAOImpl.java
#	syFinal/src/main/java/com/example/syFinal/guest/controller/GuestController.java
#	syFinal/src/main/java/com/example/syFinal/guest/controller/InfoController.java
#	syFinal/src/main/java/com/example/syFinal/guest/controller/LoginController.java
#	syFinal/src/main/java/com/example/syFinal/guest/controller/ReservController.java
#	syFinal/src/main/java/com/example/syFinal/guest/controller/ReviewController.java
#	syFinal/src/main/java/com/example/syFinal/guest/controller/WishController.java
#	syFinal/src/main/java/com/example/syFinal/guest/model/dao/GuestDAO.java
#	syFinal/src/main/java/com/example/syFinal/guest/model/dao/GuestDAOImpl.java
#	syFinal/src/main/java/com/example/syFinal/guest/model/dao/InfoDAO.java
#	syFinal/src/main/java/com/example/syFinal/guest/model/dao/InfoDAOImpl.java
#	syFinal/src/main/java/com/example/syFinal/guest/model/dao/MainDAO.java
#	syFinal/src/main/java/com/example/syFinal/guest/model/dao/MainDAOImpl.java
#	syFinal/src/main/java/com/example/syFinal/guest/model/dao/ReservDAO.java
#	syFinal/src/main/java/com/example/syFinal/guest/model/dao/ReservDAOImpl.java
#	syFinal/src/main/java/com/example/syFinal/guest/model/dao/ReviewDAO.java
#	syFinal/src/main/java/com/example/syFinal/guest/model/dao/ReviewDAOImpl.java
#	syFinal/src/main/java/com/example/syFinal/guest/model/dao/WishDAO.java
#	syFinal/src/main/java/com/example/syFinal/guest/model/dao/WishDAOImpl.java
#	syFinal/src/main/java/com/example/syFinal/guest/model/dto/GuestDTO.java
#	syFinal/src/main/java/com/example/syFinal/guest/model/dto/MainDTO.java
#	syFinal/src/main/java/com/example/syFinal/guest/model/dto/ReservDTO.java
#	syFinal/src/main/java/com/example/syFinal/guest/model/dto/ReviewDTO.java
#	syFinal/src/main/java/com/example/syFinal/host/controller/HostAccountController.java
#	syFinal/src/main/java/com/example/syFinal/host/controller/HostLoginController.java
#	syFinal/src/main/java/com/example/syFinal/host/controller/HotelController.java
#	syFinal/src/main/java/com/example/syFinal/host/controller/ReplyController.java
#	syFinal/src/main/java/com/example/syFinal/host/model/dao/HostDAO.java
#	syFinal/src/main/java/com/example/syFinal/host/model/dao/HostDAOImpl.java
#	syFinal/src/main/java/com/example/syFinal/host/model/dao/HotelDAO.java
#	syFinal/src/main/java/com/example/syFinal/host/model/dao/HotelDAOImpl.java
#	syFinal/src/main/java/com/example/syFinal/host/model/dao/ReplyDAO.java
#	syFinal/src/main/java/com/example/syFinal/host/model/dao/ReplyDAOImpl.java
#	syFinal/src/main/java/com/example/syFinal/host/model/dto/HostDTO.java
#	syFinal/src/main/java/com/example/syFinal/host/model/dto/HotelDTO.java
#	syFinal/src/main/java/com/example/syFinal/host/model/dto/HotelDetailDTO.java
#	syFinal/src/main/resources/email.properties
#	syFinal/src/main/resources/mappers/admin/admin.xml
#	syFinal/src/main/resources/mappers/admin/notice.xml
#	syFinal/src/main/resources/mappers/guest/guest.xml
#	syFinal/src/main/resources/mappers/guest/info.xml
#	syFinal/src/main/resources/mappers/guest/main.xml
#	syFinal/src/main/resources/mappers/guest/reserv.xml
#	syFinal/src/main/resources/mappers/guest/review.xml
#	syFinal/src/main/resources/mappers/guest/wish.xml
#	syFinal/src/main/resources/mappers/host/host.xml
#	syFinal/src/main/resources/mappers/host/hotel.xml
#	syFinal/src/main/resources/mappers/host/reply.xml
#	syFinal/src/main/resources/mappers/reputation.xml
#	syFinal/src/main/webapp/static/images/host/hotel/property-01.jpg

+++ syFinal/src/main/java/com/example/syFinal/host/controller/HostAccountController.java
@@ -3,6 +3,7 @@ package com.example.syFinal.host.controller;
 import java.io.File;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.UUID;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -37,10 +38,8 @@ public class HostAccountController {
 
 	@PostMapping("idCheck")
 	public Map<String, Object> idCheck(@RequestParam(name = "userEmail", defaultValue = "") String userEmail) {
-		System.out.println("===> userEmail: " + userEmail);
 		// id 중복검사
 		int checked = hostDao.idCheck(userEmail);
-		System.out.println("===> checked: " + checked);
 		Map<String, Object> data = new HashMap<>();
 		if (userEmail.equals("undefined") || checked > 0) {
 			data.put("msg", "error");
@@ -54,7 +53,6 @@ public class HostAccountController {
 	public ResponseEntity<String> pwdCheck(@PathVariable(name = "pwd") String pwd,
 			@RequestParam(name = "userEmail", defaultValue = "") String userEmail) {
 		String savedPwd = hostDao.pwdCheck(userEmail);
-		System.out.println("==> pwdCheck? " + pwd + ", ==> " + pwdEncoder.matches(pwd, savedPwd));
 		if (pwdEncoder.matches(pwd, savedPwd)) {
 			return new ResponseEntity<>("true", HttpStatus.OK);
 		} else {
@@ -66,11 +64,13 @@ public class HostAccountController {
 	@PostMapping("join")
 	public Map<String, Object> join(@RequestParam Map<String, Object> map,
 			@RequestParam(name = "profile", required = false) MultipartFile profile,
-			@RequestParam(name = "file", required = false) MultipartFile file, HttpServletRequest request) {
+			@RequestParam(name = "file", required = false) MultipartFile file,
+			@RequestParam(name = "bankbook", required = false) MultipartFile bankbook, HttpServletRequest request) {
 		ServletContext application = request.getSession().getServletContext();
 		String path = application.getRealPath("static/images/host/profile/");
 		String h_profile = "-";
 		String h_file = "-";
+		String h_bankbook = "-";
 
 		if (profile != null && !profile.isEmpty()) {
 			try {
@@ -83,18 +83,26 @@ public class HostAccountController {
 		map.put("h_profile", h_profile);
 		if (file != null && !file.isEmpty()) {
 			try {
-				file.transferTo(new File(path + h_file));
 				h_file = file.getOriginalFilename();
+				file.transferTo(new File(path + h_file));
 			} catch (Exception e) {
 				e.printStackTrace();
 			}
 		}
 		map.put("h_file", h_file);
+		if (bankbook != null && !bankbook.isEmpty()) {
+			try {
+				h_bankbook = bankbook.getOriginalFilename();
+				bankbook.transferTo(new File(path + h_bankbook));
+			} catch (Exception e) {
+				e.printStackTrace();
+			}
+		}
+		map.put("h_bankbook", h_bankbook);
 
 		// 비밀번호 암호화
 		String encodedPwd = pwdEncoder.encode((CharSequence) map.get("pwd"));
 		map.replace("pwd", encodedPwd);
-		System.out.println("===> map: " + map);
 		hostDao.insert(map);
 		Map<String, Object> data = new HashMap<>();
 		data.put("msg", "success");
@@ -104,7 +112,6 @@ public class HostAccountController {
 	@GetMapping("account/{userIdx}")
 	public Map<String, Object> getAccount(@PathVariable(name = "userIdx") int h_idx) {
 		Map<String, Object> data = hostDao.getAccount(h_idx);
-		System.out.println("===> 결과: " + data);
 		return data;
 	}
 
@@ -112,12 +119,13 @@ public class HostAccountController {
 	@PostMapping("update/{userIdx}")
 	public void updateInfo(@PathVariable(name = "userIdx") int h_idx, @RequestParam Map<String, Object> map,
 			@RequestParam(name = "profile", required = false) MultipartFile profile,
-			@RequestParam(name = "file", required = false) MultipartFile file, HttpServletRequest request) {
-		System.out.println("==> update 컨트롤러: " + map);
+			@RequestParam(name = "file", required = false) MultipartFile file,
+			@RequestParam(name = "bankbook", required = false) MultipartFile bankbook, HttpServletRequest request) {
 		ServletContext application = request.getSession().getServletContext();
 		String path = application.getRealPath("static/images/host/profile/");
 		String h_profile = "";
 		String h_file = "";
+		String h_bankbook = "";
 		if (profile != null && !profile.isEmpty()) { // 새 프로필을 업로드 한 경우,
 			try {
 				String profile1 = hostDao.getFile(h_idx, "h_profile");
@@ -151,12 +159,23 @@ public class HostAccountController {
 			h_file = hostDao.getFile(h_idx, "h_file");
 		}
 		map.put("h_file", h_file);
+		if (bankbook != null && !bankbook.isEmpty()) {
+			h_bankbook = bankbook.getOriginalFilename();
+			try {
+				bankbook.transferTo(new File(path + h_bankbook));
+				map.put("h_bankbook", h_bankbook);
+			} catch (Exception e) {
+				e.printStackTrace();
+			}
+		} else {
+			h_bankbook = hostDao.getFile(h_idx, "h_bankbook");
+		}
+		map.put("h_bankbook", h_bankbook);
 
 		// 비밀번호 암호화
 		String encodedPwd = pwdEncoder.encode(map.get("pwd").toString());
 		map.replace("pwd", encodedPwd);
 
-		System.out.println("===> map: " + map);
 		hostDao.updateInfo(map);
 	}
 
@@ -167,29 +186,43 @@ public class HostAccountController {
 			throws Exception {
 		String h_profile = hostDao.getFile(h_idx, "h_profile");
 		String h_file = hostDao.getFile(h_idx, "h_file");
-		try { // 저장된 프로필, 첨부파일이 있으면 삭제
-			if (h_profile != null && !h_profile.equals("-")) {
-				ServletContext application = request.getSession().getServletContext();
-				String path = application.getRealPath("static/images/host/profile/");
-				File profile = new File(path + h_profile);
-				if (profile.exists()) {
-					System.out.println("===> 프로필삭제? " + profile.delete());
-					profile.delete();
+		String h_bankbook = hostDao.getFile(h_idx, "h_bankbook");
+		if (hostDao.checkOrders(h_idx)) {
+			try { // 저장된 프로필, 첨부파일, 통장사본이 있으면 삭제
+				if (h_profile != null && !h_profile.equals("-")) {
+					ServletContext application = request.getSession().getServletContext();
+					String path = application.getRealPath("static/images/host/profile/");
+					File profile = new File(path + h_profile);
+					if (profile.exists()) {
+						profile.delete();
+					}
 				}
-			}
 
-			if (h_file != null && !h_file.equals("-")) {
-				ServletContext application = request.getSession().getServletContext();
-				String path = application.getRealPath("static/images/host/profile/");
-				File file = new File(path + h_file);
-				if (file.exists()) {
-					System.out.println("===> 첨부파일삭제? " + file.delete());
-					file.delete();
+				if (h_file != null && !h_file.equals("-")) {
+					ServletContext application = request.getSession().getServletContext();
+					String path = application.getRealPath("static/images/host/profile/");
+					File file = new File(path + h_file);
+					if (file.exists()) {
+						file.delete();
+					}
 				}
+
+				if (h_bankbook != null && !h_bankbook.equals("-")) {
+					ServletContext application = request.getSession().getServletContext();
+					String path = application.getRealPath("static/images/host/profile/");
+					File bankbook = new File(path + h_bankbook);
+					if (bankbook.exists()) {
+						bankbook.delete();
+					}
+				}
+				UUID uuid = UUID.randomUUID();
+				String deletedEmail = uuid.toString();
+				hostDao.deleteAccount(h_idx, deletedEmail); // 테이블에서 해당 계정정보 : 탈퇴한 회원으로 업데이트
+				return "complete";
+			} catch (Exception e) {
+				throw new Exception();
 			}
-			hostDao.deleteAccount(h_idx); // 테이블에서 해당 계정정보 delete
-			return "complete";
-		} catch (Exception e) {
+		} else {
 			throw new Exception();
 		}
 	}
@@ -201,7 +234,6 @@ public class HostAccountController {
 			hostDao.levelUp(h_idx);
 			return new ResponseEntity<>("true", HttpStatus.OK);
 		} catch (Exception e) {
-			// 에러발생
 			e.printStackTrace();
 			return new ResponseEntity<>("false", HttpStatus.BAD_REQUEST);
 		}
